longPollThreadPool can not be null
*Problem*

The long polling logic is built with the assumption that there is a thread pool for scheduling deferred reads.
So if people happens to set `numLongPollWorkerThreads` to zero or negative, a null value is passed into long poll requests which causes NPE.

*Solution*

If `numLongPollWorkerThreads` is set to zero or negative, fall back to use read thread pool. If there is no read thread pool, create a thread pool
aligned with num processors.

With this change, turn the default value of `long poll threads` to zero. so no additional threads are needed to be created if long poll feature is not used.