longPollThreadPool can not be null
Descriptions of the changes in this PR:

*Problem*

The long polling logic is built with the assumption that there is a thread pool for scheduling deferred reads.
So if people happens to set `numLongPollWorkerThreads` to zero or negative, a null value is passed into long poll requests which causes NPE.

*Solution*

If `numLongPollWorkerThreads` is set to zero or negative, fall back to use read thread pool. If there is no read thread pool, create a thread pool
aligned with num processors.

With this change, turn the default value of `long poll threads` to zero. so no additional threads are needed if long poll feature is not used (e.g. at pulsar).

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Yiming Zang <yzang2016@gmail.com>, Matteo Merli <mmerli@apache.org>

This closes #1308 from sijie/validate_number_long_poll_threads