Issue #1310: Filter out body and masterKey from ProcessorV3s
Descriptions of the changes in this PR:

In certain logs instances of ProcessorV3 (for eg WriteEntryProcessorV3) are logged.

Logging entry data and credentials details is not appropriate. Filter out
those details

(have to consider other ProcessorV3s as well and testcases to validate this change)

Master Issue: #1310

Author: cguttapalem <cguttapalem@salesforce.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>, Venkateswararao Jujjuri (JV) <None>

This closes #1311 from reddycharan/filterlogs, closes #1310