MockLedgerHandle duplicates LedgerEntries on read
Previously it was not duplicating, so the underlying ByteBuf was being
returned to the client, and this was having its reader index modified,
which meant that if it was read again, the data could not be read.

By calling duplicate() on the LedgerEntryImpl, the underlying ByteBuf
is sliced and retained, so it has a new set of indices. The retention
does nothing, as the buffer is unpooled.