Improve logging on ledger dirs monitor to avoid log flooding
Descriptions of the changes in this PR:

*Problem*

When a bookie is in readonly mode, the ledger dirs monitor will keep check the disk usage and generate tons of logs if the disk usage is unchanged. This makes debugging much difficult.

*Solution*

- Improve the logging logic in ledger dirs monitor to only log changes when disk usage is changed.
- Disable logging on checking threshold for high priority writes. Only log changes when high priority writes availability is changed.

*Result*

This reduces the logging when a bookie is outage in readonly mode.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Yiming Zang <yzang2016@gmail.com>, Matteo Merli <mmerli@apache.org>

This closes #1322 from sijie/improve_monitor_logging