Use metadata service uri rather than zkServers
Descriptions of the changes in this PR:

*Motivation*

Metadata service uri is introduced as part of BP-29. However most of the tests and tools are still assuming `zkServers`.
This causes a problem if people configures `metadataServiceUri`, some of the admin tools don't actually work.

*Solution*

This change changes bookkeeper to use `metadataServiceUri` in all the possible places and provides methods to resolve zkServers
and zkLedgersRootPath from metadata service uri. This ensures:

1) if users configure `metadataServiceUri`, all the tools would work properly.
2) if users doesn't configure `metadataServiceUri` and still use `zkServers` and `zkLedgersRootPath`, all the tools would still work.

*Changes*

The main changes are:

- deprecated direct usage of `getZkServers`, `setZkServers`, `getZkLedgersRootPath`, `setZkLedgersRootPath, `getZkAvailablePath` at AbstractConfiguration
- provide util methods `resolveZkServers` and `resolverZkLedgersRootPath` in `ZkMetadataDriverBase` for resolving zookeeper specific settings from metadata service uri
- change the tools to use `getMetadataServiceUri`
- change the tests to use `getMetadataServiceUri` to ensure code coverage

*NOTES*

There are a few places still using `getZkServers`, `setZkServers`. those are hard to change at this moment.

Related Issue:
- #1335

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1336 from sijie/fix_local_bookkeeper