Fallback to use `getZkServers` and `getZkLedgersPath` when resolving from metadata service uri
*Problme*

Dlog tests are failing because dlog is using an external zookeeper client, where both `zkServers` and `metadataServiceUri`
are not set on client configuration. It will throw NPE when trying to resolve `zkServers` and `zkLedgersRootPath` from metadata service uri.

*Solution*

Fallback to use deprecated `getZkServers` and `getZkLedgersPath` when metadata service uri is null