Fallback to use `getZkServers` and `getZkLedgersPath` when resolving from metadata service uri
Descriptions of the changes in this PR:

*Problme*

Dlog tests are failing because dlog is using an external zookeeper client, where both `zkServers` and `metadataServiceUri`
are not set on client configuration. It will throw NPE when trying to resolve `zkServers` and `zkLedgersRootPath` from metadata service uri.

*Solution*

Fallback to use deprecated `getZkServers` and `getZkLedgersPath` when metadata service uri is null

Related Issues: #1336

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1338 from sijie/fix_dlog_tests