Rev2 for Issue 791: Reduce copies in AddEntry code path
Changes in Rev 2:

- Instead of using unpooled heap ByteBuf's, we're now using pooled heap ByteBuf's. For this, we have to instantiate our own PooledByteBufAllocator(false).
- DigestManager is now passed a reference to ByteBufAllocator to use for allocations (requires change of constructors of all digest managers).
- LedgerHandle passes hint to DigestManager which ByteBufAllocator to use (based on protocol version).
- Test classes (not protocol-specific) will automatically use ByteBufAllocator for V3 protocol.

For v2 protocol, we keep using pooled direct ByteBufs as before. For v3 protocol, we switch to pooled heap ByteBuf's which are backed by an array. Added
API methods to ByteBufList allow us to get a reference to this array and wrap it into a protobuf ByteString without any copies. This reduces (for a write set
size of 3) the number of byte[] allocations and copies from 6 to 0. (Rev 1 of this change was reducing them from 6 to 1.)