[TABLE SERVICE] Dlog based checkpoint store
*Motivation*

Currently the table range stores are using local filesystem as a checkpoint store. It is okay for running as standalone mode.
But it doesn't work to run in a distributed mode. This change is introducing a dlog based checkpoint store to make sure all
the sst files of table ranges are durably checkpointed into bookkeeper itself.

*Solution*

Introduced a dlog based checkpoint store.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #1366 from sijie/dlog_checkpoint_manager