OrderedScheduler#chooseThread(key) handle null key
Descriptions of the changes in this PR:

*Motivation*

It is more convenient if OrderedScheduler#chooseThread(key) can handle null key. so the applications who is using #chooseThread don't need to worry which method to choose - `chooseThread()` vs `chooseThread(key)`.

*Solution*

in `chooseThread(key)`, fallback to randomly pickup a thread if key is null.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1372 from sijie/order_scheduler_null_key