Make bookie state manager & registration manager reusable outside of bookie
Descriptions of the changes in this PR:

*Motivation*

Since 4.6.0, we have already abstracted a pretty good interfaces around registration service and state management. Most of the classes can actually be reused for use cases that have similar service discovery and writable/readonly state transition management.

*Solution*

The change is try to cleanup `BookieStateManager` and `ZKRegistrationManager` to make them more generic and can be reused for other usage.

*Result*

Registration Interfaces and StateManager can be used outside of bookie.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1373 from sijie/refactor_bookie_state_manager