BP14 - forceLedger bookie side implementation
This is the bookie side implementation of the force() API as defined in BP-14, without the wire protocol handler.
We will introduce a new RPC which allows the client to ensure that all data sent for a  given ledger is persisted durably to the bookie, this is useful for DEFERRED_SYNC writers. In this case we are simulating a regular sync'd write to the journal choosen for the ledger.

Author: Enrico Olivelli <eolivelli@apache.org>
Author: eolivelli <eolivelli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>, JV Jujjuri <vjujjuri@salesforce.com>

This closes #1375 from eolivelli/bp14-simple-forceledger