Issue #1396: Export metrics in same http server
*Motivation*

Currently metrics provider and http admin endpoint are isolated due to the way how things are modularized.
This requires two ports to be used, one is for metrics provider to expose metrics, while the other one is used by
http admin endpoint for expose admin endpoints.

When implementing a bookkeeper operator on k8s, it become inconvinient on managing those ports and do health checks.
It would be good to allow export metrics in same http admin endpoint if it is enabled.

*Solution*

- Introduce a method `writeAllMetrics` in stats provider to allow stats provider exporting metrics.
- Add a `MetricsService` in bookie http service and export stats provider's metrics under `/metrics` endpoint.

This fixes #1396