Provide a vagrantfile to launch a linux vm for running integration tests
*Motivation*

Currently all bookkeeper's integration tests and bc tests are docker based and written
using arquillian framework. Due to the network issues in docker and the way how bookies
advertise themselves in the cluster, we can *ONLY* run integration tests in a linux
environment. It is inconvinient to people who is using macOS as the development environment.

*Solution*

This PR is providing a Vagrantfile to lanuch a dev vm that contains all the tools required
for running integration tests and bc tests. So people can actually debug and test integration
tests locally using a linux vm.