Fix TestSmoke#testTailingReads
Descriptions of the changes in this PR:

*Motivation*

TestSmoke#testTailingReads is testing the tailing behavior at bookkeeper. However the current testing logic is a bit
problematic on checking whether the reader has caught up with the expected last add confirmed. so the read logic break
earlier before reading the expectedLastConfirmedEntry.

*Solution*

Fix the checking logic on breaking the read loop.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>

This closes #1402 from sijie/fix_integration_tests