Provide zookeeper startup script
Descriptions of the changes in this PR:

*Motivation*

ZooKeeper is a dependency of bookkeeper shipped along with bookkeeper dist package. However we never provide any script to startup zookeeper in the bookkeeper binary package and never test the zookeeper version shipped along with bookkeeper

*Solution*

- add zookeeper in `bin/bookeeper` scripts to start zookeeper using bookkeeper scripts
- refactor bookkeeper docker scripts to start bookie/zookeeper/standalone
- use zk adminPort for health check, since 4letters commands are explicitly disabled at 3.5

*Tests*

Existing integration tests test the docker scripts and the bash script changes to start zookeeper.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>

This closes #1416 from sijie/package_zookeeper