Allow bookie to start if failed to load extra server components
*Motivation*

`extraServerComponents` provides the flexibility to allow bookie to start with extra server components.
It acts as a plugin mechanism to extend the functionality of bookies. However it is okay to allow bookie
start up when failed to load extra server components.

*Solution*

Catch the exception and log it when failed to load extra server components.