Fix dev/check-binary-license on macOS
Descriptions of the changes in this PR:

*Motivation*

`dev/check-binary-license` is used for checking whether LICENSE files are updated to reflect the dependencies included in the distribution packages.
However this script doesn't work on macOS, which makes the development on macOS inconvinient.

*Solution*

The change removes `--wildcards` from tar command. The option only works on linux machines.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #1418 from sijie/fix_check_license_script_master