Build stream modules for integration tests
Descriptions of the changes in this PR:

*Motivation*

apache/bookkeeper#1422 adds table service as part of integration tests. so we need to build stream modules
in order to run integration tests for apache/bookkeeper#1422

*Solution*

Add "-Dstream" in the maven commands for `precommit-integrationtests` CI job.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #1424 from sijie/compile_stream_module