Allow bookie to start if failed to load extra server components
Descriptions of the changes in this PR:

*Motivation*

`extraServerComponents` provides the flexibility to allow bookie to start with extra server components.
It acts as a plugin mechanism to extend the functionality of bookies. However it is okay to allow bookie
start up when failed to load extra server components.

This is a change for allowing starting table service as extra components in bookies.

*Solution*

Catch the exception and log it when failed to load extra server components.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1420 from sijie/allow_start_bookie