BP-14 WriteFlag DEFERRED_SYNC Client Side Implementation
Introduce implementation of WriteFlag#DEFERRED_SYNC on LedgerHandle (purely client-side).

Tests are only on the client-side (with Mockito), there is no implementation on the bookie side. Most significant tests will come with the force() API, at this point we can only test that we have not broken the LAC advance mechanism.

Author: Enrico Olivelli <eolivelli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>, Venkateswararao Jujjuri (JV)

This closes #853 from eolivelli/bp14-deferred-force-write-client-side