[TABLE SERVICE] start table service as an extra component of bookie
Descriptions of the changes in this PR:

*Motivation*

table service was built over bookkeeper ledgers. it is an extension to bookies' functionalities, so it is much convenient to start the service as one additional component with bookie, just like how we start `AutoRecovery` along with bookie.

*Solution*

- include `stream/server` module as part of bookkeeper server/all binary distributions
- abstract `StorageServer` to allow controlling whether to start bookie or not
- provide a ServerLifecycleComponent of `StorageServer`, so it can be used as an extra component of bookie

*Tests*

- improve the integration tests to start table service as part of bookie containers
- move `LocationClientTest` to container based integration tests

Master Issue: #1205

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1422 from sijie/start_table_service