BP-33: Move releasing official docker images out of main repo
*Motivation*

Current bookkeeper docker images are auto-built by apache docker account. However it becomes problematic in the release process:

Docker autobuild uses release tag for labeling the versions for docker images. But the Dockerfile can only be updated after a release is successfully made. So we have to retag a release after a release, in order to update Dockerfile to build the docker images.

Master Issue: #1449