Allow CI to publish a snapshot version with gitsha information
Descriptions of the changes in this PR:

*Motivation*

Currently bookkeeper publishes a snapshot release nightly to maven artifacts. It provides a convenient solution for people to try out new bookkeeper releases as soon as the changes land in master branch. However sometime it is also problematic when you reference this snapshot version in other projects, subsequent changes in bookkeeper might potentially break those projects.

*Solution*

Update the CI to provide a mechanism to publish a snapshot version with gitsha information. so it can publish a snapshot version that is pinned to the given gitsha. so the projects can safely use a snapshot version for development without worrying bookkeeper changes break it.

*NOTES*

The mechanism is a manual operation. It should be used when it is needed.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1454 from sijie/publish_snapshot_with_gitsha