[table service] provide unified service uri for resolving service endpoints
*Motivation*

Currently there are multiple settings in `StorageClientSettings` on configuring how to resolve
service endpoints to find the services. It is a bit inconvinient and confusing

*Solution*

Provide a class `ServiceURI` for parsing service uri to retrieve common informations including:

- serviceName
- serviceInfos
- serviceUser
- serviceHosts
- servicePath

It extends the java.net.URI and does parsing and validation in the same place. Then it can be used
by the grpc name resolver to resolve service endpoints.

Also update dlog to use it for resolving its namespace uri.