BP-33: Move releasing official docker images out of main repo
Descriptions of the changes in this PR:

*Motivation*

Current bookkeeper docker images are auto-built by apache docker account. However it becomes problematic in the release process:

Docker autobuild uses release tag for labeling the versions for docker images. But the Dockerfile can only be updated after a release is successfully made. So we have to retag a release after a release, in order to update Dockerfile to build the docker images.

Master Issue: #1449

Author: Sijie Guo <sijie@apache.org>

Reviewers: Ivan Kelly <ivank@apache.org>, Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #1450 from sijie/official_docker_images