Improve merge script to handle PRs based on branches
*Motivation*

Merge script doesn't handle well on handling PRs based on branches. So committers ended up
using merge button to merge PRs. so we lose a lot of informations such as descriptions, reviewers
in the commit message.

*Solution*

This PR fixes the problems on handling merging PRs based on branches.

This allows us merging #1469