[BOOKIE] [DBLEDGERSTORAGE] fix illegal reference count exception on filling read cache
Descriptions of the changes in this PR:

*Problem*

ByteBuf is released at a finally block. DbLedgerStorage doesn't have to release it on return. If it does so, that will cause
double-release and it will throw IllegalReferenceCountException.

*Solution*

Remove `ByteBuf.release` at return.

Related Issue: #1487

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>, Matteo Merli <mmerli@apache.org>

This closes #1488 from sijie/fix_double_release_on_db_ledgerstorage