Issue 1476: LedgerEntry is recycled twice at ReadLastConfirmedAndEntryOp
*Problem*

There are two flags on checking whether a request is completed. That is being misued for two different branches,
one is when the request is completed with advanced lac but no entry piggybacked, the other one is when the request
is completed with adavanced lac with an entry piggybacked. When this happen, it will cause a request being completed
twice and the entry buffer is recycled twice.

*Solution*

Remove direct usage of submitCallback and use completeRequest instead.

Add a unit test for reproduce the issue and ensure the fix address the problem.