ISSUE #1490: BookieJournalForceTest is flaky
Descriptions of the changes in this PR:

### Motivation

The test cases in BookieJournalForceTest check if the forceWriteQueue is empty after everything is flushed.
However if `adaptiveGroupWrites` is enabled, the forceWrite thread will insert a marker entry into force write queue,
this will fail the forceWriteQueue length checking.

### Changes

Disable `adaptiveGroupWrites` in BookieJournalForceTest

Master Issue: #1490

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1512 from sijie/fix_journalforcetest, closes #1490